修复bug

+++ pkg/runner/options.go
@@ -32,6 +32,7 @@ type Options struct {
 	ErrUseLastResponse bool                `json:"err_use_last_response,omitempty"`
 	Csv                bool                `json:"csv,omitempty"`
 	ClearResume        bool                `json:"clear_resume,omitempty"`
+	Version            bool                `json:"version,omitempty"`
 }
 
 func ParserOptions() *Options {
@@ -57,6 +58,7 @@ func ParserOptions() *Options {
 		set.BoolVarP(&options.Silent, "silent", "sl", false, "只输出状态码为200"),
 		set.BoolVarP(&options.EnableProgressBar, "progressbar", "pb", false, "启用进度条"),
 		set.BoolVar(&options.Skip404And302, "skip", false, "跳过404、302输出默认跳过"),
+		set.BoolVarP(&options.Version, "version", "v", false, "输出版本"),
 	)
 	set.CreateGroup("config", "配置",
 		set.IntVarP(&options.Retries, "retries", "rs", 3, "重试3次"),
@@ -73,9 +75,14 @@ func ParserOptions() *Options {
 		set.BoolVar(&options.ClearResume, "clear", false, "清理历史任务"),
 	)
 	_ = set.Parse()
+	if options.Version {
+		showBanner()
+		gologger.Print().Msgf("pathScan version: %s", Version)
+		os.Exit(0)
+	}
 	if options.ClearResume {
 		_ = os.RemoveAll(DefaultResumeFolderPath())
-
+		gologger.Print().Msgf("clear success: %s", DefaultResumeFolderPath())
 		os.Exit(0)
 	}
